backend terminado melldroga

+++ src/main/java/com/puce/redsocial/Controllers/ComentariosController.java
@@ -2,11 +2,12 @@ package com.puce.redsocial.Controllers;
 
 import com.puce.redsocial.Entitys.Comentarios;
 import com.puce.redsocial.Services.ComentariosService;
-import jakarta.annotation.security.RolesAllowed;
+import com.puce.redsocial.Services.PublicacionService;
+import com.puce.redsocial.Services.UsuarioService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
+import java.nio.file.AccessDeniedException;
 import java.util.Optional;
 
 @RestController
@@ -14,25 +15,38 @@ import java.util.Optional;
 public class ComentariosController {
     @Autowired
     private ComentariosService comentariosServ;
+    @Autowired
+    private PublicacionService publicacionServ;
+    @Autowired
+    private UsuarioService usuarioServ;
 
     @GetMapping
-    public List<Comentarios> getAllComentarios() {
+    public Iterable<Comentarios> getAllComentarios() {
         return comentariosServ.getAllComentarios();
     }
 
     @GetMapping("/{id}")
-    public Optional<Comentarios> getComentarioById(@PathVariable Long id) {
+    public Optional<Comentarios> getComentarioById(@PathVariable Integer id) {
         return comentariosServ.getComentarioById(id);
     }
 
-    @PostMapping
-    public Comentarios createComentario(@RequestBody Comentarios comentarios) {
+    @RequestMapping(value="/create", method = RequestMethod.POST, produces = "application/json")
+    public Comentarios createComentario(@RequestBody Comentarios comentarios) throws AccessDeniedException {
+        if(publicacionServ.getPublicacionById(comentarios.getPublicacion()).isEmpty()){
+            throw new AccessDeniedException("Publicación no existe");
+        }
+        if(usuarioServ.getUserById(comentarios.getUsuario()).isEmpty()){
+            throw new AccessDeniedException("Usuario no existe");
+        }
         return comentariosServ.createComentario(comentarios);
     }
 
-    @DeleteMapping("/{id}")
-    @RolesAllowed("ADMINISTRADOR")
-    public void deleteComentario(@PathVariable Long id) {
+    @DeleteMapping("/delete/{id}")
+    public void deleteComentario(@PathVariable Integer id, @RequestHeader Integer userid) throws AccessDeniedException {
+        Comentarios cometarios = comentariosServ.getComentarioById(id).get();
+        if (!cometarios.getUsuario().equals(userid)) {
+            throw new AccessDeniedException("Acción denegada");
+        }
         comentariosServ.deleteComentario(id);
     }
 }
